[type:refactor] fix lost logic bug
