+++ shenyu-plugin/shenyu-plugin-global/pom.xml
@@ -28,7 +28,7 @@
     <dependencies>
         <dependency>
             <groupId>org.apache.shenyu</groupId>
-            <artifactId>shenyu-plugin-api</artifactId>
+            <artifactId>shenyu-plugin-base</artifactId>
             <version>${project.version}</version>
         </dependency>
         <dependency>
