+++ shenyu-plugin/shenyu-plugin-global/src/main/java/org/apache/shenyu/plugin/global/subsciber/MetaDataCacheSubscriber.java
@@ -17,7 +17,7 @@
 
 package org.apache.shenyu.plugin.global.subsciber;
 
-import org.apache.shenyu.plugin.global.cache.MetaDataCache;
+import org.apache.shenyu.plugin.base.cache.MetaDataCache;
 import org.apache.shenyu.common.dto.MetaData;
 import org.apache.shenyu.sync.data.api.MetaDataSubscriber;
