+++ shenyu-plugin/shenyu-plugin-global/src/test/java/org/apache/shenyu/plugin/global/subsciber/MetaDataCacheSubscriberTest.java
@@ -18,7 +18,7 @@
 package org.apache.shenyu.plugin.global.subsciber;
 
 import org.apache.shenyu.common.dto.MetaData;
-import org.apache.shenyu.plugin.global.cache.MetaDataCache;
+import org.apache.shenyu.plugin.base.cache.MetaDataCache;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
